- renamed the getting started samples to overview
